+++ src/Utils/SslFd.cpp
@@ -34,14 +34,14 @@ int SslFd::GetError(int shakeHandRet)
     return ::SSL_get_error(sslHandle_, shakeHandRet);
 }
 
-int SslFd::Write(const std::string& writeBuf)
+ssize_t SslFd::Write(const std::string& writeBuf)
 {
     return ::SSL_write(sslHandle_,
                        reinterpret_cast<const void*>(writeBuf.data()),
                        static_cast<int>(writeBuf.size()));
 }
 
-int SslFd::Read(std::string& readBuf)
+ssize_t SslFd::Read(std::string& readBuf)
 {
     return ::SSL_read(sslHandle_,
                       reinterpret_cast<void*>(readBuf.data()),
